Propagate graph update list to UR (#17019)
Update the `urCommandBufferUpdateKernelLaunchExp` API for updating
commands in a command-buffer to take a list of commands.

The current API operates on a single command, this means that the
SYCL-Graph `update(std::vector<nodes>)` API needs to serialize the list
into N calls to the UR API. Given that both OpenCL
`clUpdateMutableCommandsKHR` and Level-Zero
`zeCommandListUpdateMutableCommandsExp` can operate on a list of
commands, this serialization at the UR layer of the stack introduces
extra host API calls.

This PR improves the `urCommandBufferUpdateKernelLaunchExp` API so that
a list of commands is passed all the way from SYCL to the native backend
API.

As highlighted in
https://github.com/oneapi-src/unified-runtime/issues/2671 this patch
requires the handle translation auto generated code to be able to handle
a list of structs, which is not currently the case. This is PR includes
a API specific temporary workaround in the mako file which will unblock
this PR until a more permanent solution is completed.

---------

Co-authored-by: Ross Brunton <ross@codeplay.com>