Fix some tests that are broken when run with multiple cuda devices available. (#17216)
Also removes a test and adds known failures where appropriate (typically
where the test only runs when multiple devices are available so the skip
doesn't affect behaviour of single-device runs).

The test removed is `cudaUrContextCreateTest.ActiveContext`. This test
seems to be testing the assumption that a `urQueueCreate` followed by
`urMemBufferCreate` will set the active cuda context to the one
associated with the context passed to those calls. Neither of these
calls set the active context, this may have changed at some point as the
test dates back to a PI unit test. The test currently passes as long as
only one device is available because a previous `urDeviceGetInfo` sets
the active context to the one associated with that device, which is
inevitably the same as the one associated with the UR context used in
the test. Since the test is based on a faulty assumption about the
adapter I think we can just delete it.