Fix cupti library dynamic loading (#17272)
This patch fixes loading the cupti library on Linux to rely on the
dynamic linker to figure out where the library is.

It also disable cupti tracing on Windows, as far as I know this is only
used on Linux at the moment since the tracing tools don't support
Windows. Additionally there are some potential security issues of
loading the library just by name on Windows, so we just leave Windows
support out for now.

This is following up on the discussions on
https://github.com/oneapi-src/unified-runtime/pull/1070, statically
linking the cupti library was also ruled out as it makes the adapter
library go from around 600KiB to around 32MiB.